Move interfaces to model.ts
